refactor(config): Change the creation of GitHubConfigFileProvider
Use a factory function in the companion object that resolves the
properties from the configuration. This simplifies the constructor,
since the handling of the configuration is done by the factory
function.

Store only the base URL for GitHub requests; the other information is
not needed.

Move the logging of the configuration from the factory class to the
new factory function.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>