feat(workers): Resolve template files in the reporter worker
Introduce a prefix for template files to be used in the reporter job
configuration. Pre-process the options in the job configuration to
download all referenced template files. Pass the processed options to
the reporters.

Note: The reporter worker is the first worker that references
configuration files in its options. Therefore, the functionality to
download these files is added here. It may later be extracted and
generalized to be available to other workers as well.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>