test(workers): Improve resolutions in `OrtTestData`
Make the issue and vulnerability resolutions match the issue and
vulnerability contained in the result. Change the rule violation
resolution to match any rule violation because the test data does
currently not contain a rule violation. Also align the messages of the
resolutions in the repository configuration and the resolved
configuration.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>