refactor(workers): Remove a default parameter for `readConfigFile()`
Production code typically always has to provide the resolved
configuration context when downloading configuration files. So, remove
the default value of `null` for this parameter, since it poses the
risk that the default context is used instead of the one which has
been resolved by the config worker at the beginning of the ORT run.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>