fix(workers): Pass the correct configuration context
In ConfigWorker, make sure that the `WorkerContext` passed to the
validation script contains the correct resolved configuration context.
The configuration context is obtained from the `OrtRun` wrapped by the
worker context; but this instance does not yet contain the resolved
configuration context, which is stored at the end of the worker
execution.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>