feat(config): Extend the `ConfigManager.downloadFile()` function
The function now allows overriding the name of the downloaded file. If
no target file name is specified, it derives the file name from the
configuration path.

If multiple configuration files are downloaded that reference each
other, the file names are relevant. Therefore, the previous approach
to generate a temporary file name was insufficient.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>