feat(worker): Extend download support in WorkerContext
The functions to download configuration files now support the same
arguments as `ConfigManager.downloadFile()`; so a target directory and
a target file name can be specified. The automatic removal of
downloaded files has been dropped, since the same can be achieved
easier with the support for temporary directories.

The new functionality is needed by the Reporter worker to download not
only template files but also arbitrary assets like fonts and images.
ReporterRunner has been adapted to the change of the method
signatures.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>