feat(model): Support downloading assets in ReporterJobConfiguration
Introduce a new ReporterAsset class that describes a file or a
directory to be downloaded. Add lists of such objects to
ReporterJobConfiguration. The idea is that the reporter worker should
download these files before starting with the report generation. That
way asset files like images or fonts can be made available to
reporters.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>