feat(reporter): Support a placeholder for the working directory
For some reporters it is necessary to reference from the options to
paths relative to the current working directory. For instance, the PDF
reporter needs the path where custom fonts are stored.

To support this, introduce a placeholder that is looked up in the
reporter options and is replaced by the current (temporary) working
directory.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>