feat: Upgrade ORT and configuration classes
Temporarily switch back to the JitPack releases of ORT to be able to
upgrade to the current revision which includes important changes for the
scanner and advisor configurations. Plugins for both tools are now
configured via maps of options and secrets, and the implementation
specific configuration classes for advice providers have been removed.

This requires several adaptions which would have been difficult to split
into individual commits, therefore all changes have been implemented in
a single commit. The most important changes are:

* The `AdvisorJobConfiguration` and `ScannerJobConfiguration` classes
  have been adapted to the new configuration model.
* The database tables for the old configuration model have been removed
  and tables for the new model have been added. To reduce the effort,
  there is no migration for the old configuration data.
* The advisor worker now instantiates the advice providers based on the
  job configuration instead of the `application.conf`.

Note that the secret options are not yet transformed to the actual
secret values from the secret storage. This will be implemented in a
follow-up commit.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>