fix(Config): Change issue message in case of validation script errors
The message of the `OrtIssue` that was generated in case of an error
to execute the parameter validation script (which is typically caused
by compilation failures in the script) was too long and caused
exceptions on database inserts. To prevent this, use a short error
message in the issue and write more details to the log.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>