refactor(jobmonitor): Move common functionality to JobHandler
The functionality to send notifications for failed jobs and only to
delete failed jobs if this was successful was implemented by both
JobMonitor and Reaper. Move this functionality to JobHandler to reduce
code duplication. This simplifies JobMonitor and Reaper, and now
JobHandler is the only component that needs to deal with
FailedJobNotifier.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>