fix(jobmonitor): Prevent duplicate notifications for failed job
When the Reaper detected and deleted a failed job, JobMonitor was
triggered as well because it received an update event for the affected
job. So two notifications about a job failure were sent to
Orchestrator.

Prevent this by keeping a list of the jobs that have been processed
recently; so that duplicate processing of jobs can be detected.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>