feat(core): Run sync of Keycloak roles in background
Previously, the synchronization was done in blocking mode, and the API
was responding only when it was done. Since it can take a long time,
this was problematic.

Note that the solution implemented now is just a quick fix for this
problem. Running the synchronization at the startup of the core
component is still not ideal.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>