fix(core): Fix the path for the Keycloak API secret
The Keycloak API secret was wrongly read from the path
`keycloakApiSecret` which resolved to `keycloak.keycloakApiSecret` in
the `application.conf` and to `keycloakApiSecret` in the secret storage.
This was wrong, because the value in the `application.conf` is called
`keycloak.apiSecret`.

Fix this by not using the `getConfig` function to get the nested
configuration for Keycloak and instead using the full paths for
accessing the Keycloak configuration options, similar to how it is
implemented for the database configuration.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>