fix(dao): Fix potential unique constraint violation exceptions
When populating the intermediate table for many-to-many relations a
check for duplicate entries was missing. This could cause exceptions
due to constraint violations.

To fix this, introduce a new helper function that performs a mapping
from model objects to data objects and filters out duplicates. Use
this function wherever intermediate tables are involved.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>