fix(dao): Make parsing of license finding start lines more robust
LicenseFindingCurationDao stores the start lines of license findings
as comma-separated string in the database, but exposes them as list of
integers. Handle invalid number values during this conversion.

Note that in normal operation the table should not contain invalid
values. It is, however, possible that incorrect data is inserted
manually or by other tools.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>