feat(scanner): Associate package provenances with scanner runs
Extend the `OrtServerPackageProvenanceStorage` to associate package
provenances with the current scanner run. This solves the issue that the
scanner run returned by the `get` functions in `DaoScannerRunRepository`
contains scan results for packages which were ignored by the ORT
scanner. For example, the ORT scanner ignores excluded packages if the
`skipExcluded` option is enabled.

To be able to associate package provenances with scanner run it was
necessary to remove the unique constraints from the
`package_provenances` table as they prevented keeping the results for
older scanner runs. Accordingly, the read functions of the
`OrtServerPackageProvenanceStorage` now return the latest entry instead
of the single one.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>