feat(scanner): Create scanner run in DB before running the scanner
Create the scanner run entry in the database before starting the ORT
scanner. This makes it possible to associate data in the provenance and
scan result storages with the scanner run.

To be able to create the scanner run in the database before the result
data is available, several properties have been made nullable and the
function to create scanner runs was split into a create function that
creates an empty scanner run, and a function to finalize the scanner run
once the data is available.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>