feat(scanner): Associate package provenances and nested provenances
Since b7a3d45 multiple package provenances for the same root provenance
are not overwritten anymore but kept for older runs. Apply the same
change for nested provenances to ensure that the resolution results for
older scans are kept and the results for those scans are still correct,
even if a new nested provenance resolution was performed for the same
package provenance (this happens for example, if the nested provenance
contains sub-repositories which are referenced by branch names, as the
revisions of those branches could have changed since the last
resolution).

To be able to store multiple nested provenances for the same root
provenance it was necessary to remove the unqiue constraint from the
`nested_provenance` table. Accordingly, the read function of the
`OrtServerNestedProvenanceStorage` now returns the latest entry instead
of the single one.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>