chore(logging): Slightly change the logging pattern
When using a logging aggregation system, it is typically helpful to
have a more structured format (with key value-pairs) for log
statements to enable improved query facilities. For now, assign a
label to the log level, which is needed for a level-based filtering.
In the future, more labels can be added to support upcoming use cases.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>