refactor(scanner): Switch the implemented storage read function
Implement the `read` function that takes a `provenance` and a `matcher`
instead of the one that only takes a `provenance`. This will be required
by an upcoming change which needs access to the scan results after
filtering them with the `matcher`.

The previously implemented function is never used by ORT itself so it
can safely throw an exception if it is called.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>