feat(scanner): Associate scan results with scanner runs
Extend the `OrtServerScanResultStorage` to associate scan results with
the current scanner run. This solves the issue that it was previously
difficult to figure out which scan results for a package belong to a
scanner run, because multiple results from different scanners and
versions could be stored.

To be able to associate scan results with scanner runs it was necessary
to remove the unique constraints from the `scan_results` table as they
prevented to keep results for older scanner runs for scanners which
always re-run, like FossID.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>