fix(orchestrator): Do no set `startedAt` timestamp for jobs
Do not set the `startedAt` timestamp when the orchestrator schedules
jobs. The property is supposed to contain the timestamp when the actual
processing of the job started, not when the job was scheduled for
execution. Otherwise any queue times would be contained when calculating
the duration of a job which can make it hard to interpret the result. If
the timestamp when a job was scheduled is required, a separate
`scheduledAt` property should be introduced.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>