feat(reporter): Use `PluginConfiguration` for reporter options
Since some reporters may require credentials to interact with external
services, the `Reporter` interface expects a `PluginConfiguration` to
be passed in. Support this now in ORT Server, too, by changing the
reporter job configuration accordingly.

In the reporter worker, handle the configurations for the single
reporters correctly and also resolve secrets if needed.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>