fix(dao): Fix problems with indexes
Indexes on text columns that can have an arbitrary length are
problematic, since an exception is thrown when inserting texts that
are too long. Prevent this for a number of tables.

For the tables `vulnerabilities` and `defects`, indexes for value
fields were missing. Create those to speed up `getOrPut` operations,
but exclude unlimited text columns.

For the `ort_issues` table, drop the unique constraint, since the
de-duplication logic of issues is going to be removed. The required
code changes for this are done in a follow-up commit.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>