ci: Prepare for enabling GitHub merge queues
Prepare the GitHub action configurations for enabling merge queues. For
details see [1] and [2].

Merge queues allow to ensure that all PRs are tested on top of the main
branch without requiring the author to manually rebase the PR. This
helps to prevent issues that occur when the checks for individual PRs
pass but the combination of both PRs breaks a check. An example for this
is when two PRs introduce different database migrations with the same
version.

[1]: https://docs.github.com/en/actions/using-workflows/events-that-trigger-workflows#merge_group
[2]: https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/managing-a-merge-queue

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>