feat(analyzer): Add a custom PackageCurationProvider implementation
This implementation expects curation files in a folder structure
following certain naming conventions. It differs from ORT's
`FilePackageCurationProvider` in that it does not load all curation
files initially, but resolves the files for packages on demand. For
the intended usage scenario - the provider is used for a single
Analyzer run -, this skips the overhead of initialization, which can
be significant for a large number of curations or if files are loaded
from a slow network share.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>