fix(scanner): Extend `PackageProvenanceCache`
In order to correctly assign provenances with a subpath to their
nested provenance, more information has to be stored in the cache. The
API needs to be extended, so that assignments can be made no matter in
which order the data arrives. This means that also put() operations
have to return data, because when new information is added to the
cache, it can be the case that assignments of nested provenances can
now be performed.

The ORT Server-specific provenance storage implementations still need
to be extended to make use of this new functionality.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>