feat(transport): Extend the `MessageHeader` class by properties
The idea is that a transport implementation can treat messages in a
special way based on the properties they contain. One use case is to
support different tool versions: A message could for instance have a
property `javaVersion` with a value `17`. The transport implementation
can then route the message to an endpoint that fulfills this
requirement.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>