feat(model): Change the parameters in job configurations
Instead of having separate maps with parameters for the different
jobs, define only a single map in `JobConfigurations`. For the
intended use case - to support a simplified, customized API on top of
the job configurations API -, this representation is easier to use for
end users. The previous approach with multiple maps would require at
least some knowledge about the distribution of tasks to different
workers.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>