feat(kubernetes): Support setting resource requirements
Add new properties to `KubernetesSenderConfig` to set limits and
requests for the "cpu" and "memory" resources. When creating a job,
evaluate these properties, so that Kubernetes is informed about the
pod's requirements.

The properties support variable interpolation. This can be used to
adjust the values dynamically. For instance, if a large repository is
processed, the limits could be increased.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>