feat(build): Use the current Java version to set the jvmTarget
This is a preparation to support container images for different Java
versions. The idea is to create those images by building parts of the
codebase (namely the Analyzer worker) on the desired target Java
version. Note that ORT uses an analogous strategy to determine the
target JVM when it is built.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>