feat(build): Make the base image tag configurable for jib
This is needed to integrate other base images built for different Java
versions. Currently, this feature is only used by the Analyzer worker.
To align the worker builds, support this mechanism for the other
workers as well. The only exception is the Advisor worker, which
currently does not use a custom base image, but directly extends the
Eclipse Temurin image.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>