feat(workers): Make infrastructure services optional in configuration
So far, the section with infrastructure services was mandatory in the
`.ort.env` file and in `EnvironmentConfig`. There are, however, use
cases which do not require that services are defined; for instance,
when referencing services from the product or organization or to
define environment variables. Therefore, make this property optional.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>