chore(core): Change a sort parameter
When requesting the runs of a repository sort by the run index instead
of the creation date. This should typically yield the same results.

This change is actually done to fix a flaky integration test which
creates multiple runs and then queries them. Since the runs have been
created almost at the same time, the sort order based on the creation
time was not reliable.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>