refactor(api): Move mapping code to separate module
Move the code to map between API and business model to a separate Gradle
module. This allows to remove the business model dependency from the API
module which prevents that business model classes are accidentally used
in the API model.

While at it, also fix that Konform is an api dependency, not an
implementation dependency of the API module.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>