build: Use `:utils:test` in all modules
Add `:utils:test` as a test dependency in all modules to make the
shared test configuration available to all tests. Currently this
includes the defined labels and the deactivation of Kotest classpath
scanning.

Ideally, this would be implemented centrally using a Gradle convention
plugin, but this would have required a bigger refactoring of the Gradle
setup.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>