feat(job-monitor): Add another notification function
Add a function to `FailedJobNotifier` to send notifications about lost
jobs. This is different from the existing function, since there is no
Kubernetes job available.

Currently, the function sends a `WorkerError` message, analogous to
failed jobs. This may be changed in the future, so that Orchestrator
could distinguish these cases.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>