feat(api): Allow CORS
Allow CORS[1] from all hosts to allow frontend development. This is
required due to the UI and API being served on different origins, or ports
if the case of the development environment.
This should be replaced with a configurable CORS address in the future to
allow only configuring the frontend address as the allowed origin.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS

Signed-off-by: Mikko Murto <mikko.murto@doubleopen.org>