fix(core): Do not require authentication to download reports by tokens
Although there was no check for permissions, the route to download a
report using a token still required authentication. Fix this by
extracting the code (together with the API documentation and
integration tests) to a new function, which is called outside the
`authenticate` block in Routing.kt. This has the additional benefit
that the endpoint which does not require authentication is separated
from other routes on runs.

While at it, rename the path. Use the prefix `downloads`, which is
generic, so that this mechanism could be extended to other artifacts
in the future.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>