refactor(workers): Move handling of ORT config to a separate class
Introduce a new `WorkerOrtConfig` class that is now responsible for
setting up the environment related to global ORT configuration
options. This new class allows configuring a proper environment in
scenarios where a worker context cannot be used. An upcoming commit
will use this for the Analyzer worker if the process needs to be
forked.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>