fix(validation): Make `update` use the same validation as `create`
The change in aa53753 was a bit too eager in that it inlined validation
separately for `create` and `update` endpoints for the same entity,
whereas these should use exactly the same validation. To solve this,
move the validation constants to the respective main entity classes and
use them for both `create` and `update` calls from there.

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>