feat(compose): Remove ActiveMQ and related profiles
Remove ActiveMQ from the Docker Compose configuration and also remove
the now obsolete `rabbitmq` profile. This simplifies the usage of
Compose as no profile has to be provided anymore. Also, all other
services are configured to use RabbitMQ, so activating the `artemis`
profile would not have worked out of the box anyway.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>