ci: Add a GitHub workflow to build and publish the Docker images
Add a workflow that builds all Docker images in a matrix job and
publishes them to ghcr.io. The worker base images are published
separately, this allows to use them when building locally.

All Docker base image builds use the registry cache to improve
performance. The Jib builds are not cached, this might later be added by
using the GitHub Actions cache.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>