fix(snippets): Remove the unique constraint on the snippets table
In production, there are some errors when inserting snippets in the
table, because the column "additional_data" is too large for the index
related to the unique constraint on this table's columns.
This index is also growing and is currently taking more place than the
snippet data itself.
However, this column `additional_data` is also containing the FossID
snippet id, which is unique for every FossID scan. Therefore, this
unique constraint does not make sense, since there will never be
duplicates entries. Hence, it can be removed.

Signed-off-by: Nicolas Nobelis <nicolas.nobelis@bosch.com>