feat(ui): Edit, delete Organizations
Clicking the organization link takes the user to the details page, where
the organization can be edited or deleted. The breadcrumb in the header
shows the name of currently opened organization.

A response from the back-end is shown in form of a toast message when
creating, editing or deleting organizations.

In case of a successful operation, a general success message is shown in
the UI.  In case of request failure, the API failure response with
details is shown instead.

At this point of the UI project, it is probably most beneficial to see
the exact technical cause of the API error in the UI, as it is written
to api/v1 spec, so the cause of the error is also shown as-is to the
user in the toast message.

Later on, the message could be changed into something more user
friendly.

(This implements the "UD" part of CRUD.)

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>