refactor(orchestrator): Rework the scheduling of worker jobs
Instead of having hard-coded conditional logic, use a more declarative
approach that defines the jobs to run and their dependencies. A
generic scheduling function obtains the current status of the ORT run
and determines which jobs can run now.

This new approach makes it easier to figure out which worker jobs are
when scheduled during an ORT run. It is also more flexible with
regard to changes in the scheduling algorithm.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>