refactor(orchestrator): Rework handling of Config worker results
The Config worker is a bit special, since it does not spawn dedicated
worker jobs. With some more refactorings, it is now possible to reuse
the major part of the default scheduling when handling result messages
from this worker. This makes it possible to remove some more redundant
code from the `Orchestrator` class.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>