feat(ui): Create, list ORT runs
This commit adds the ability to create (minimal) ORT runs, list the runs
with their updated statuses and navigate to the run details page. Note
that all of this is currently implemented with minimal configuration, so
an empty config is passed to the run, resulting only the ORT Analyzer
being run. At this phase, the UI is thus implemented in a
proof-of-concept manner.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>