feat(orchestrator): Support running workers if the run failed
Add a flag to `WorkerScheduleInfo` that controls whether the worker
can run even if the ORT run is already in failed state. Set this flag
for the Reporter worker, so that always reports are generated.

Do some more refactoring in `OrchestratorTest` to reduce code
duplication, which makes it easier to test the execution of the
Reporter worker.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>