feat(ui): Insert some run parameters to analyzer
Add "allow dynamic versions" and "skip excluded" run parameters to
the Analyzer, to show that job configs can be passed to the different
jobs of an ORT run successfully.

Also add some report formats to the Reporter; so far, "ortresult" and
"WebApp" have been tested to generate the expected output files.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>