feat(ui): Implement download links for the result files of an ORT Run
Add download links for all the result files to the ORT run page.

Currently the files are downloaded by using fetch and the direct API
endpoint.  With upcoming changes to the back-end, the file download can
probably be simplified, by relying on some features in modern browsers
that support file downloading.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>