refactor(notifier): Add `NotifierOrtResultGenerator` class
This class is able to create the `OrtResult` object the Notifier
operates on, including the additional labels available to Notifier
scripts. In the existing implementation, the code to construct this
result was distributed, making it hard to see which information was
actually added to it. The new class is also a good place to document
the Notifier-specific labels.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>