feat(notifier): Support labels for querying the status of worker jobs
These labels can be used from a Notifier script to determine which
worker jobs were executed during an ORT run and whether they were
successful or not. Based on this information, it is possible to
generate different notifications for successful or failed ORT runs,
also taking the affected worker steps into account.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>