feat(core): Filter organizations based on permissions
Change the `/organizations` endpoint to only require authentication
instead of the superuser role, and to return only those organizations
for which the user has `OrganizationPermission.READ`.

As the permissions are managed by Keycloak, the database does not know
about them. Therefore the implementation fetches the whole list of
organizations from the database and filters them afterward. For a very
long list of organizations this might not perform well, so if required,
a different solution has to be found later.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>