fix(core): Fix OpenAPI representation of nullable properties
When using kotlinx.serialization with ktor-swagger-ui, the
json-schema-serialization [1] library is used for generating JSON
schemas [2].

For nullable values, this library generates invalid schemas which causes
issues with consumer of the generated OpenAPI specification. Introduce a
post-processing step to fix the schema.

Both the original json-schema-serialization library [3] and the used
fork [1] are inactive, so we should not expect a fix for this issue.

A similar issue was fixed in ktor-swagger-ui [4], but the fix does not
apply when using json-schema-serialization.

Fixes #165.

[1]: https://github.com/tillersystems/json-schema-serialization
[2]: https://github.com/SMILEY4/ktor-swagger-ui/wiki/Kotlinx-Multiplatform-Encoder-Example
[3]: https://github.com/Ricky12Awesome/json-schema-serialization
[4]: https://github.com/SMILEY4/ktor-swagger-ui/issues/53

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>