fix(gradle): Correctly parse blank `dockerBaseBuildArgs`
Splitting an empty string does not return an empty list, but a list with
an empty string as the single entry, which results in `--build-arg`
having an empty argument.

This only surfaced when running the task under WSL with Docker Desktop
for Windows, apparently because empty arguments are handled differently
there.

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>