build: Add a new utils module for YAML serialization
This is a work-around for a dependency conflict caused by different
versions of the SnakeYaml Engine library required by the Kaml
serialization library and jRuby. The module acts as a wrapper on top
of Kaml and provides a function to deserialize YAML documents. It uses
the Gradle shadow plugin [1] to relocate the conflicting SnakeYaml
packages to another package structure. This removes the conflicting
version from the classpath.

[1]: https://imperceptiblethoughts.com/shadow/

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>