chore(kubernetes): Drop a misleading log output
If no variables for annotations are specified, there is always a log
line printed about invalid variable names. This is caused by the usage
of the `split()` function which returns a non-empty array for an empty
string. (This had no further impact, because the empty variable name
was just ignored.)

Handle this case correctly, so that the confusing log line disappears.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>