fix(dao): Update repository URLs with user information
This commit adds a new migration file to replace the user information
in the URLs with 'xxx' in the repositories tables. If there are
multiple entries with the same URL after the replacement, it will
append a numerical counter (e.g. 'xxx2', 'xxx3'). In addition, tests
are added to ensure the migration works properly.

Signed-off-by: Onur Demirci <onur.demirci@bosch.com>