feat(model): Support credentials types for infrastructure services
So far, for infrastructure services it could be configured whether
their credentials should be listed in the `.netrc` file or not. It
turned out that this is not sufficient for specific repositories.

Therefore, make the concept more generic by introducing a new
`CredentialsType` enumeration. This is going to determine, in which
configuration files the service is listed. For now, support the type
"Git credentials" in addition to "netrc". In an upcoming commit,
functionality will be added to generate `.git-credentials` files.

Unfortunately, this is a rather huge change which affects all layers.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>