feat(ui): Use coloring for ORT Run statuses
To improve the UX, use colored badges to indicate the statuses of the
ORT runs in the repository and run views.

As described in [1], fast refresh doesn't work, if a file exports
something else than the component created in it, which leads to eslint
errors. To fix this, the additional export from the badge component is
moved to a separate file.

[1]: https://github.com/ArnaudBarre/eslint-plugin-react-refresh

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>