refactor(ui): Regenerate query client
OpenAPI query client library update to v1 (PR #157) introduced major
changes to the structure and functionality of the query client.
Moreover, PR #181 introduced a fix to OpenAPI generation of nullable
values, which brought major changes to the Swagger JSON.

This commit regenerates the query client from scratch to adapt to these
changes. The command used for regenerating is:

    pnpm codegen

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>