refactor(ui): Prepend unused query keys with an underscore
There is a bug [1] in the codegen tool that causes it to generate unused
query keys for requests with no parameters. This commit renames the
unused query keys to start with an underscore '_' to pass the lint
checks for unused variables.

NOTE: This is not a lasting solution, as every regeneration of the query
client will revert these changes.

[1]: https://github.com/7nohe/openapi-react-query-codegen/issues/112

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>