refactor(ui): Remove `any` type from error toasts
After the recent update to openapi-react-query-codegen[1] version 1, the
generated error body types are now `unknown` instead of `any`[2]. This
requires casting the error body to a type that contains the `message` and
`cause`. Casting is generally not a good practice, but as `message` and
`cause` can both be `undefined`, casting and accessing them is safe.

[1]: https://github.com/7nohe/openapi-react-query-codegen
[2]: https://github.com/eclipse-apoapsis/ort-server/pull/224/files#diff-4321bfeea0ab44bb44c6bc087b447bfb839f938ba3dac1c532c1757fa8d61d1dR8

Signed-off-by: Mikko Murto <mikko.murto@doubleopen.org>