fix(reporter): Add Git to the Reporter base image
This is required for the special source code bundle reporter which may
download artifacts to be included into the bundle from their VCS
location.

Note that this is a quick fix. It is also not complete, since not all
version control systems supported by ORT are added. For the long run,
a more generic solution needs to be found.

Signed-off-by: Oliver Heger <oliver.heger@bosch.io>