feat(ui): Enable/disable jobs from an ORT run
Make the ORT run creation form cleaner by using expandable accordion
components for the individual jobs of a run. By clicking on an accordion
header, the corresponding job parameters are shown for the user to edit.

Also add a switch to each job to enable or disable jobs from the run.
When a job is disabled, no corresponding job config object is sent to
the created ORT run, in effect excluding that job from the run.
By default, only Analyzer and Reporter jobs are enabled.

Signed-off-by: Jyrki Keisala <jyrki.keisala@doubleopen.org>