build: Add a Gradle task to generate the OpenAPI specification
Add the Gradle task `:core:generateOpenApiSpec` to write the OpenAPI
specification to `core/build/openapi/openapi.json`. This requires
starting the Ktor server as the specification is only generated at
runtime.

The easiest way to implement that was to create a fake test as this can
make use of the already configured test application and does not require
a custom configuration of the Ktor application. Another benefit is that
the code is not added to the runtime classpath.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>