build(ui): Generate the API code from the Gradle output
Change `generate:api` to use the `openapi.json` generated by the Gradle
task `generateOpenApiSpec` instead of taking it from a running server.

This way the only precondition for generating the API code is to run a
Gradle task instead of having to start the Docker Compose environment.

As the new script does not make use of environment variables anymore it
now also works on Windows.

Fixes #231.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>