build(ui): Remove generate API client from Git to generate it on the fly
Instead of committing the generated API client to Git, generate it on
the fly when building the ui. This ensures that the client is always in
sync with the used version of `@7nohe/openapi-react-query-codegen`.

With this change it is now mandatory to run the Gradle task
`:core:generateOpenApiSpec` to be able to build the UI, so adapt the
README and the GitHub workflows accordingly.

The `prettier` command ignores paths from `.gitignore` by default. As we
still want the generated code to be formatted, apply the workaround from
[1].

[1]: https://github.com/prettier/prettier/issues/15438

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>