feat(core): Allow multiple hosts for CORS
Rename the config property `ktor.cors.allowedHost` to
`ktor.cors.allowedHosts` and let it take a comma separated list of
hosts.

This will be needed for the Docker Compose setup in the future, to allow
to run the UI Docker image in parallel to the UI in dev mode at
different ports. Also add the host that will be used by the UI Docker
image to the default value.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>