feat(compose): Add the Keycloak client `ort-server-ui`
Add a Keycloak client to be used by the UI service in Docker Compose. It
is setup in the same way as `ort-server-ui-dev` except for using
`http://localhost:8082` for the root and home URL.

Having both clients allows for using the UI in Docker Compose and in dev
mode in parallel.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.com>